update: change navigation and info tab
